requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
numpy>=1.23.5
tensorflow>=2.13.0
opencv-python-headless>=4.8.0
Pillow>=10.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
aiofiles>=23.2.0
pydantic>=2.5.0
